+++ packages/orq-rc/src/orq_ai_sdk/models/createpromptop.py
@@ -1950,6 +1950,7 @@ CreatePromptProvider = Literal[
     "moonshotai",
     "zai",
     "minimax",
+    "xai",
     "alibaba",
     "slack",
 ]