+++ packages/orq-rc/src/orq_ai_sdk/models/deploymentinvokeop.py
@@ -85,6 +85,7 @@ Provider = Literal[
     "moonshotai",
     "zai",
     "minimax",
+    "xai",
     "alibaba",
     "slack",
 ]