+++ packages/orq-rc/src/orq_ai_sdk/models/deploymentsop.py
@@ -623,6 +623,7 @@ DeploymentsProvider = Literal[
     "moonshotai",
     "zai",
     "minimax",
+    "xai",
     "alibaba",
     "slack",
 ]