+++ packages/orq-rc/src/orq_ai_sdk/models/deploymentstreamop.py
@@ -2211,6 +2211,7 @@ DeploymentStreamProvider = Literal[
     "moonshotai",
     "zai",
     "minimax",
+    "xai",
     "alibaba",
     "slack",
 ]