+++ packages/orq-rc/src/orq_ai_sdk/models/getallpromptsop.py
@@ -530,6 +530,7 @@ GetAllPromptsProvider = Literal[
     "moonshotai",
     "zai",
     "minimax",
+    "xai",
     "alibaba",
     "slack",
 ]