+++ packages/orq-rc/src/orq_ai_sdk/models/getonepromptop.py
@@ -493,6 +493,7 @@ GetOnePromptProvider = Literal[
     "moonshotai",
     "zai",
     "minimax",
+    "xai",
     "alibaba",
     "slack",
 ]