+++ packages/orq-rc/src/orq_ai_sdk/models/getpromptversionop.py
@@ -529,6 +529,7 @@ GetPromptVersionProvider = Literal[
     "moonshotai",
     "zai",
     "minimax",
+    "xai",
     "alibaba",
     "slack",
 ]