+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -148,7 +148,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KJFCKVCCHV8B1G13ZQ7W6NS0"
+        "01KJMVZPRBS3C7JDCW3QA11GYF"
     )
     r"""The unique identifier of the data source"""