+++ packages/orq-rc/src/orq_ai_sdk/models/listpromptversionsop.py
@@ -539,6 +539,7 @@ ListPromptVersionsProvider = Literal[
     "moonshotai",
     "zai",
     "minimax",
+    "xai",
     "alibaba",
     "slack",
 ]