+++ packages/orq-rc/src/orq_ai_sdk/models/updatememorystoreop.py
@@ -104,6 +104,7 @@ UpdateMemoryStoreProvider = Literal[
     "moonshotai",
     "zai",
     "minimax",
+    "xai",
     "alibaba",
     "slack",
 ]