+++ packages/orq-rc/RELEASES.md
@@ -6219,3 +6219,13 @@ Based on:
 - [python v4.5.0-rc.16] packages/orq-rc
 ### Releases
 - [PyPI v4.5.0-rc.16] https://pypi.org/project/orq-ai-sdk/4.5.0-rc.16 - packages/orq-rc
+
+## 2026-03-01 14:16:39
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.736.2 (2.845.15) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.5.0-rc.17] packages/orq-rc
+### Releases
+- [PyPI v4.5.0-rc.17] https://pypi.org/project/orq-ai-sdk/4.5.0-rc.17 - packages/orq-rc